Add End slime entity
This commit ports the End slime entity. I also discovered a minor bug with
the spawn eggs I added yesterday, namely that middle-clicking the mod
entities in creative mod does not give you the corresponding spawn egg.
This is because the SpawnEggItem.EGGS map is never filled with the
mod entities, because items are registered before entity types in Forge.
I will have to come up with a solution to this at some point.