Fix middle click mod entity to give spawn egg
Since items are registered before entitytypes in Forge, this means that
the mod spawn eggs will not be added to the EGGS map in the SpawnEggItem
class, which in turn means that middle clicking entities in creative
mode will not give the player the appropriate spawn egg. This was fixed by
subscribing to the FMLCommonSetupEvent (where we know what all
entitytypes have been registered), and use reflection to add all the mod spawn
eggs to the EGGS map.