Add Cubozoa entity
This commit ports the Cubozoa entity, including the related items:
gelatine, sweet berry jelly, and shadow berry jelly. Note however: I did
not use data generators to generate the recipes, AND there is a small
discrepancy from the fabric version, since the recipes takes a water
bottle, but the vanilla recipe class can not distinguish between items
that only differ in nbt. This means that any potion can be used in the
jelly recipes. I might fix this later, but I consider it a low priority
problem.