Add RespawnObeliskBlock, raw amber and amber gem
This is the first small step towards porting the Amberland biome.
Amberland introduces the amber ore, which drops raw amber, which are
crafted into amber gems, which are used as 'currency' for the
RespawnObeliskBlock. Note however that this commit does NOT include the
amber ore itself, but only the items and block mentioned in the commit
title. Also note that no crafting recipes are included, which means
that the block and items can only be obtained in creative mode
as of right now.