Add crafting table variants for mod trees
This commit adds a crafting table block for each type of mod wood,
similar to the composter variants from my last commit. Note that a
custom crafting block class was needed (ModCraftingTableBlock), since
the vanilla crafting table block uses a crafting container that specifically
checks that the block clicked is the vanilla crafting table, which of
course would fail.