Add chest variants for mod trees
This commit adds chest variants for the different mod trees. Note that
the mod chest ITEMS currently have the vanilla chest texture, due to the
fact that the ChestItemTileEntityRenderer calls the
EndChestTileEntityRenderer with a tile entity without a block in the
world, and will thus default to the vanilla chest texture. Will be fixed
in future commit.