Fix method overriding in SilkMothEntity
It turns out that I forgot to add the @Override annotation to some of
the methods when I ported the silk moth entity, which of course meant
that the method was not overriding anything at all, due to different
name mapping in forge and fabric. This small issue has now been fixed.